pyproject.toml
[project]
name = "polarsteps-album-generator"
version = "0.1.0"
description = "Generate photo albums from Polarsteps trip data"
requires-python = ">=3.13"
authors = [{ name = "Itay Raveh", email = "itay.raveh@proton.me" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
keywords = [
    "polarsteps",
    "photo-album",
    "travel",
    "html-generator",
    "pdf-generator",
]
dependencies = [
    "aiohttp",
    "Pillow",
    "pydantic>2",
    "pydantic-settings>2",
    "jinja2",
    "more-itertools",
    "lxml",
    "geopy",
    "geopandas",
    "shapely",
    "rich",
    "tenacity",
    "diskcache",
    "pyproj",
    "nicegui",
    "opencv-python-headless",
]

[dependency-groups]
dev = [
    "ruff",
    "ty",
    "basedpyright",
    "deptry",
    "djlint",
    "taskipy",
    "lxml-stubs",
    "types-shapely",
    "types-geopandas",
]

[build-system]
requires = ["uv_build>=0.9.23,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "psagen"
module-root = "src"

[tool.taskipy.tasks]
format = "ruff format ."
check = "ruff check . --fix --unsafe-fixes"
types = "ty check && basedpyright"
deps = "deptry ."
html = "djlint static/*.jinja static/templates/*.jinja --lint --reformat"
lint = "task format && task check && task types && task deps && task html"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["ALL"]

ignore = [
    "D203",    # 1-blank-line-before-class (conflicts with D211)
    "D213",    # Multi-line docstring summary at second line (conflicts with D212)
    "COM812",  # Missing trailing comma (conflicts with formatter)
    "ISC001",  # Single line implicit string concatenation (conflicts with formatter)
    "FIX002",  # Fails if "TODO" is present (blocks development too hard)
    "TD003",   # Missing issue link for TODO (too bureaucratic)
    "EM101",   # Exception must not use a string literal
    "EM102",   # Exception must not use an f-string literal
    "PLR2004", # Magic numbers - many are acceptable (HTTP codes, common values)
    "TRY003",  # Long exception messages - acceptable for user-facing errors
    "TRY400",
    "ANN401",  # Any types - acceptable for generic cache functions
    "S324",    # MD5 hash - acceptable for non-cryptographic use (color selection)
    "D100",    # Missing docstring in public module
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
    "D104",    # Missing docstring in public package
    "D105",    # Missing docstring in magic method
    "D106",    # Missing docstring in public nested class
    "D107",    # Missing docstring in __init__
    "D401",
    "RUF017",
]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.mccabe]
max-complexity = 8

[tool.ty]

[tool.pyright]
reportUnusedCallResult = false
reportPrivateUsage = false
reportImplicitOverride = false
reportUnannotatedClassAttribute = false

[tool.deptry.per_rule_ignores]
# DEP002 = "Unused dependency"
# We ignore these because they are CLI tools, not libraries we import.
DEP002 = [
    "pre-commit",
    "taskipy",
    "playwright",
    "ruff",
    "mypy",
    "deptry",
    "djlint",
]

[tool.djlint]
profile = "jinja"
# H006: img tags missing height/width (often impossible in dynamic templates)
# H031: meta keywords missing (outdated SEO rule)
ignore = "H006,H031"
